+++ src/Components/Auth/AuthLogin.js
@@ -12,8 +12,9 @@ const AuthLogin = () => {
 
   const history = useHistory();
 
+  // Continue button sends users to workout tab after logging in
   const buttonHandler = () => {
-    history.push("/User/workouts");
+    history.push("/workouts");
   };
 
   // flags in the state to watch for add/remove updates