+++ src/Components/Auth/AuthRegister.js
@@ -15,8 +15,9 @@ const AuthRegister = () => {
 
   const history = useHistory();
 
+  // Continue button sends users to workout tab after registering
   const buttonHandler = () => {
-    history.push("/User/workouts");
+    history.push("/workouts");
   };
 
   // flags in the state to watch for add/remove updates