+++ src/Components/Components.js
@@ -1,8 +1,10 @@
 import React from "react";
 import Login from "./Auth/AuthLogin";
 import Register from "./Auth/AuthRegister";
-import ExerciseAuthModule from "./Exercise/ExerciseAuth.js";
-import WorkoutAuthModule from "./Workout/WorkoutAuth.js";
+import ExerciseModule from "./Exercise/Exercise";
+import ExerciseAuthModule from "./Exercise/ExerciseAuth";
+import WorkoutModule from "./Workout/Workout";
+import WorkoutAuthModule from "./Workout/WorkoutAuth";
 
 import Menubar from "./Menubar/Menubar.js";
 import AuthModule from "./Auth/Auth.js";
@@ -22,8 +24,11 @@ const Components = () => {
           {/* Routing added here, will be expanded with future components */}
           {/* New tabs must also be added with new routing */}
           <Route path="/" exact component={AuthModule} />
-          <Route path="/:username/workouts" component={WorkoutAuthModule} />
-          <Route path="/:username/exercises" component={ExerciseAuthModule} />
+          {/* These routes are used after authentication, sends user to their personal app display */}
+          <Route path="/:username/workouts" component={WorkoutModule} />
+          <Route path="/:username/exercises" component={ExerciseModule} />
+          <Route path="/workouts" component={WorkoutAuthModule} />
+          <Route path="/exercises" component={ExerciseAuthModule} />
           <Route path="/login" component={Login} />
           <Route path="/register" component={Register} />
           <Redirect to="/" />